+++ cfgov/v1/forms.py
@@ -30,6 +30,7 @@ class FilterDateField(forms.DateField):
                 pass
         return value
 
+
 class PDFFilterDateField(forms.DateField):
     def clean(self, value):
         if value:
@@ -38,6 +39,7 @@ class PDFFilterDateField(forms.DateField):
             except Exception as e:
                 pass
         return value
+
 
 class FilterCheckboxList(forms.CharField):
     def validate(self, value):
@@ -49,12 +51,18 @@ class FilterCheckboxList(forms.CharField):
 
 
 class CalenderPDFFilterForm(forms.Form):
-    filter_calendar = FilterCheckboxList(label='Calendar',
-        error_messages=ERROR_MESSAGES['CHECKBOX_ERRORS'])
-    filter_range_date_gte = PDFFilterDateField(required=False,
-        error_messages=ERROR_MESSAGES['DATE_ERRORS'])
-    filter_range_date_lte = PDFFilterDateField(required=False,
-        error_messages=ERROR_MESSAGES['DATE_ERRORS'])
+    filter_calendar = FilterCheckboxList(
+        label='Calendar',
+        error_messages=ERROR_MESSAGES['CHECKBOX_ERRORS']
+    )
+    filter_range_date_gte = PDFFilterDateField(
+        required=False,
+        error_messages=ERROR_MESSAGES['DATE_ERRORS']
+    )
+    filter_range_date_lte = PDFFilterDateField(
+        required=False,
+        error_messages=ERROR_MESSAGES['DATE_ERRORS']
+    )
 
     def __init__(self, *args, **kwargs):
         kwargs['error_class'] = FilterErrorList
@@ -66,12 +74,14 @@ class CalenderPDFFilterForm(forms.Form):
     def clean(self):
         cleaned_data = super(CalenderPDFFilterForm, self).clean()
         from_date_empty = 'filter_range_date_gte' in cleaned_data and \
-                          cleaned_data['filter_range_date_gte']  == None
+                          cleaned_data['filter_range_date_gte'] is None
         to_date_empty = 'filter_range_date_lte' in cleaned_data and \
-                        cleaned_data['filter_range_date_lte'] == None
+                        cleaned_data['filter_range_date_lte'] is None
 
         if from_date_empty and to_date_empty:
-            raise forms.ValidationError(ERROR_MESSAGES['DATE_ERRORS']['one_required'])
+            raise forms.ValidationError(
+                ERROR_MESSAGES['DATE_ERRORS']['one_required']
+            )
         return cleaned_data
 
 
@@ -115,7 +125,6 @@ class FilterableListForm(forms.Form):
         self.set_topics(page_ids)
         self.set_authors(page_ids)
 
-
     def clean_categories(self):
         """ This is a (hopefully) temporary solution for dealing w/ the fact
         that we show Blog and Reports as options for filtering, but
@@ -140,9 +149,13 @@ class FilterableListForm(forms.Form):
         return categories
 
     def base_query(self):
-        base_query = AbstractFilterPage.objects.live_shared(hostname=self.hostname)
+        base_query = AbstractFilterPage.objects.live_shared(
+            hostname=self.hostname
+        )
         if self.parent:
-            base_query = base_query.filter(CFGOVPage.objects.child_of_q(self.parent))
+            base_query = base_query.filter(
+                CFGOVPage.objects.child_of_q(self.parent)
+            )
             logger.info('Filtering by parent {}'.format(self.parent))
         return base_query
 
@@ -159,7 +172,9 @@ class FilterableListForm(forms.Form):
 
     # Populate Topics' choices
     def set_topics(self, page_ids):
-        tags = Tag.objects.filter(v1_cfgovtaggedpages_items__content_object__id__in=page_ids).values_list('slug', 'name')
+        tags = Tag.objects.filter(
+            v1_cfgovtaggedpages_items__content_object__id__in=page_ids
+            ).values_list('slug', 'name')
 
         options = self.prepare_options(arr=tags)
         most = options[:3]
@@ -171,7 +186,9 @@ class FilterableListForm(forms.Form):
 
     # Populate Authors' choices
     def set_authors(self, page_ids):
-        authors = Tag.objects.filter(v1_cfgovauthoredpages_items__content_object__id__in=page_ids).values_list('slug', 'name')
+        authors = Tag.objects.filter(
+            v1_cfgovauthoredpages_items__content_object__id__in=page_ids
+            ).values_list('slug', 'name')
         options = self.prepare_options(arr=authors)
 
         self.fields['authors'].choices = options
@@ -254,6 +271,7 @@ class EventArchiveFilterForm(FilterableListForm):
 
 
 class FeedbackForm(forms.ModelForm):
+    """For feedback modules that simply ask 'Was this page helfpul?'"""
     class Meta:
         model = Feedback
         fields = ['is_helpful', 'comment']
@@ -261,3 +279,30 @@ class FeedbackForm(forms.ModelForm):
     def __init__(self, *args, **kwargs):
         super(FeedbackForm, self).__init__(*args, **kwargs)
         self.fields['is_helpful'].required = True
+
+
+class ReferredFeedbackForm(forms.ModelForm):
+    """For feedback modules that need to capture the referring page"""
+    class Meta:
+        model = Feedback
+        fields = ['referrer', 'comment']
+
+    def __init__(self, *args, **kwargs):
+        super(ReferredFeedbackForm, self).__init__(*args, **kwargs)
+        self.fields['comment'].required = True
+
+
+class SuggestionFeedbackForm(forms.ModelForm):
+    """For feedback modules seeking content suggestions"""
+
+    class Meta:
+        model = Feedback
+        fields = ['referrer',
+                  'comment',
+                  'expect_to_buy',
+                  'currently_own',
+                  'email']
+
+    def __init__(self, *args, **kwargs):
+        super(SuggestionFeedbackForm, self).__init__(*args, **kwargs)
+        self.fields['comment'].required = True